+++ tools/src/metadata-generator.ts
@@ -49,7 +49,8 @@ export default class DXComponentMetadataGenerator {
 
             logger('Generate metadata for ' + widgetName);
 
-            let isTranscludedContent = widget['IsTranscludedContent'],
+            let isTranscludedContent = widget.IsTranscludedContent,
+                isExtension = widget.IsExtensionComponent || false,
                 dasherizedWidgetName = inflector.dasherize(inflector.underscore(widgetName)),
                 outputFilePath = path.join(config.outputFolderPath, dasherizedWidgetName.substr('dx-'.length) + '.json'),
                 events = [],
@@ -96,6 +97,8 @@ export default class DXComponentMetadataGenerator {
                 className: inflector.classify(widgetName),
                 widgetName: widgetName,
                 isTranscludedContent: isTranscludedContent,
+                isExtension: isExtension,
+                baseClass: isExtension ? 'DxComponentExtension' : 'DxComponent',
                 selector: dasherizedWidgetName,
                 events: allEvents,
                 properties: properties,